[MOVE/#22] component 파일 이동
