+++ core-ui/build.gradle
@@ -58,4 +58,7 @@ dependencies {
     debugImplementation 'androidx.compose.ui:ui-tooling'
     debugImplementation 'androidx.compose.ui:ui-test-manifest'
 
+    // coil
+    implementation "io.coil-kt:coil-compose:1.4.0"
+
 }
