+++ feature/src/main/java/com/sopt/feature/auth/signup/SignUpScreen.kt
@@ -1,6 +1,5 @@
 package com.sopt.feature.auth.signup
 
-import android.util.Log
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -12,7 +11,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -24,8 +22,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.navigation.NavHostController
-import com.sopt.component.textfield.TextFieldWithTitle
 import com.sopt.feature.R
+import com.sopt.feature.component.textfield.TextFieldWithTitle
 import com.sopt.ui.extension.toast
 
 @Composable
